requirements.txt
numpy
pandas
sklearn
colorlog
requests
scipy
python-dotenv